chat component separated
